Fix DataLoader
DataLoader로 데이터가 한꺼번에 주입될 수 있도록 Fix
Application

+++ src/main/java/com/kernel360/hackerthon/studyup/member/repository/PeerReviewRepository.java
@@ -9,10 +9,10 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface PeerReviewRepository extends JpaRepository<PeerReview, BigInteger> {
+public interface PeerReviewRepository extends JpaRepository<PeerReview, Long> {
     List<PeerReview> findAll();
 
-    Optional<PeerReview> findById(BigInteger id);
+    Optional<PeerReview> findById(Long id);
 
     PeerReview save(PeerReview peerReview);
