+++ src/main/java/com/kernel360/hackerthon/studyup/study/service/StudyService.java
@@ -24,7 +24,7 @@ public interface StudyService {
      * @return
      * DTO 객체 새로 생성 ( 스터디 정보 + 스터디에 속한 멤버 + 스터디 리뷰 + 기술 스택 + 스택 별점)
      */
-    StudyGroupDetailDTO getStudyById(BigInteger studyId);
+    StudyGroupDetailDTO getStudyById(Long studyId);
 
     /**
      * 스터디 그룹 개설
